fix datapoint deserialzation

+++ carbonj.service/src/main/java/com/demandware/carbonj/service/db/points/DataPointArchiveRocksDB.java
@@ -217,7 +217,7 @@ class DataPointArchiveRocksDB
 
             try
             {
-                int mId = DataPointRecord.toMetricId( key );
+                long mId = DataPointRecord.toMetricId( key );
                 int ts = DataPointRecord.toTimestamp( key );
                 msg = "Failed to remove record. db [" + dbName + "], key [" + mId + ":" + ts + "]";
             }
